googleai: start moving PaLM client into googleai
