+++ llms/googleai/googleai_llm.go
@@ -1,7 +1,5 @@
 // package googleai implements a langchaingo provider for Google AI LLMs.
-// See https://ai.google.dev/ for more details and documetnation.
-//
-// nolint: lll
+// See https://ai.google.dev/ for more details.
 package googleai
 
 import (
