+++ llms/googleai/palm/internal/vertexaiclient/palm_client.go
@@ -1,4 +1,4 @@
-package vertexaiclient
+package palmclient
 
 import (
 	"context"
