+++ llms/googleai/palm/vertexai_palm_llm.go
@@ -1,4 +1,4 @@
-package vertexai
+package palm
 
 import (
 	"context"
@@ -6,7 +6,7 @@ import (
 
 	"github.com/tmc/langchaingo/callbacks"
 	"github.com/tmc/langchaingo/llms"
-	"github.com/tmc/langchaingo/llms/vertexai/internal/vertexaiclient"
+	"github.com/tmc/langchaingo/llms/googleai/palm/internal/vertexaiclient"
 )
 
 var (
