+++ llms/googleai/palm/vertexai_palm_llm_option.go
@@ -1,4 +1,4 @@
-package vertexai
+package palm
 
 import (
 	"net/http"
